Skipping transpilation won't discourage minifying
Here and elsewhere, we've seen arguments that this proposal will encourage users to skip minifying their code. (I think the argument is that since users are required to transpile, that will also encourage them to minify.)

The FAQ has an entry on this, but I wanted to elaborate on it, by pointing out that no popular transpiler minifies by default. Minification always requires a separate, non-default step.